feat: #3-#4 - download data from data.calgary.ca
This code utilizes a Socrata client from sodapy

The pandas DataFrame will be persisted on first run,
and then loaded from the local file `pathways.pkl` if it exists.
This is to ensure we don't hit the data.calgary.ca
throttle limit on free accounts.

This DataFrame is not in the same form as those produced
by esri2gpd. It may require transformation to get at the
appropriate coordinates.